refac(zk): move stringifiers to `zk/plonk/halo2/stringifiers`
