+++ tachyon/zk/base/BUILD.bazel
@@ -18,29 +18,11 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
-    name = "field_stringifier",
-    hdrs = ["field_stringifier.h"],
-    deps = [
-        "//tachyon/base/strings:rust_stringifier",
-        "//tachyon/math/finite_fields:prime_field_base",
-    ],
-)
-
-tachyon_cc_library(
     name = "point_set",
     hdrs = ["point_set.h"],
     deps = [
         "//tachyon/base:ref",
         "//tachyon/base/containers:container_util",
-    ],
-)
-
-tachyon_cc_library(
-    name = "point_stringifier",
-    hdrs = ["point_stringifier.h"],
-    deps = [
-        ":field_stringifier",
-        "//tachyon/math/elliptic_curves:points",
     ],
 )
