+++ tachyon/zk/plonk/halo2/pinned_constraint_system.h
@@ -11,12 +11,12 @@
 #include <string>
 #include <vector>
 
-#include "tachyon/zk/lookup/lookup_argument_stringifier.h"
-#include "tachyon/zk/plonk/base/phase_stringifier.h"
 #include "tachyon/zk/plonk/constraint_system/constraint_system.h"
-#include "tachyon/zk/plonk/constraint_system/query_stringifier.h"
 #include "tachyon/zk/plonk/halo2/pinned_gates.h"
-#include "tachyon/zk/plonk/permutation/permutation_argument_stringifier.h"
+#include "tachyon/zk/plonk/halo2/stringifiers/lookup_argument_stringifier.h"
+#include "tachyon/zk/plonk/halo2/stringifiers/permutation_argument_stringifier.h"
+#include "tachyon/zk/plonk/halo2/stringifiers/phase_stringifier.h"
+#include "tachyon/zk/plonk/halo2/stringifiers/query_stringifier.h"
 
 namespace tachyon {
 namespace zk::halo2 {
