+++ tachyon/zk/plonk/halo2/stringifiers/challenge_stringifier.h
@@ -4,8 +4,8 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CHALLENGE_STRINGIFIER_H_
-#define TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CHALLENGE_STRINGIFIER_H_
+#ifndef TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_CHALLENGE_STRINGIFIER_H_
+#define TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_CHALLENGE_STRINGIFIER_H_
 
 #include <ostream>
 
@@ -28,4 +28,4 @@ class RustDebugStringifier<zk::Challenge> {
 
 }  // namespace tachyon::base::internal
 
-#endif  // TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CHALLENGE_STRINGIFIER_H_
+#endif  // TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_CHALLENGE_STRINGIFIER_H_
