+++ tachyon/zk/plonk/halo2/stringifiers/column_key_stringifier.h
@@ -4,14 +4,14 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_BASE_COLUMN_KEY_STRINGIFIER_H_
-#define TACHYON_ZK_PLONK_BASE_COLUMN_KEY_STRINGIFIER_H_
+#ifndef TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_COLUMN_KEY_STRINGIFIER_H_
+#define TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_COLUMN_KEY_STRINGIFIER_H_
 
 #include <ostream>
 
 #include "tachyon/base/strings/rust_stringifier.h"
 #include "tachyon/zk/plonk/base/column_key.h"
-#include "tachyon/zk/plonk/base/column_type_stringifier.h"
+#include "tachyon/zk/plonk/halo2/stringifiers/column_type_stringifier.h"
 
 namespace tachyon::base::internal {
 
@@ -59,4 +59,4 @@ class RustDebugStringifier<zk::ColumnKey<C>> {
 
 }  // namespace tachyon::base::internal
 
-#endif  // TACHYON_ZK_PLONK_BASE_COLUMN_KEY_STRINGIFIER_H_
+#endif  // TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_COLUMN_KEY_STRINGIFIER_H_
