+++ tachyon/zk/plonk/halo2/stringifiers/column_type_stringifier.h
@@ -4,25 +4,25 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_BASE_PHASE_STRINGIFIER_H_
-#define TACHYON_ZK_PLONK_BASE_PHASE_STRINGIFIER_H_
+#ifndef TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_COLUMN_TYPE_STRINGIFIER_H_
+#define TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_COLUMN_TYPE_STRINGIFIER_H_
 
 #include <ostream>
 
 #include "tachyon/base/strings/rust_stringifier.h"
-#include "tachyon/zk/plonk/base/phase.h"
+#include "tachyon/zk/plonk/base/column_type.h"
 
 namespace tachyon::base::internal {
 
 template <>
-class RustDebugStringifier<zk::Phase> {
+class RustDebugStringifier<zk::ColumnType> {
  public:
   static std::ostream& AppendToStream(std::ostream& os, RustFormatter& fmt,
-                                      zk::Phase phase) {
-    return os << fmt.DebugTuple("Phase").Field(int{phase.value()}).Finish();
+                                      zk::ColumnType type) {
+    return os << zk::ColumnTypeToString(type);
   }
 };
 
 }  // namespace tachyon::base::internal
 
-#endif  // TACHYON_ZK_PLONK_BASE_PHASE_STRINGIFIER_H_
+#endif  // TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_COLUMN_TYPE_STRINGIFIER_H_
