+++ tachyon/zk/plonk/halo2/stringifiers/permutation_argument_stringifier.h
@@ -4,13 +4,13 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_PERMUTATION_PERMUTATION_ARGUMENT_STRINGIFIER_H_
-#define TACHYON_ZK_PLONK_PERMUTATION_PERMUTATION_ARGUMENT_STRINGIFIER_H_
+#ifndef TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_PERMUTATION_PERMUTATION_ARGUMENT_STRINGIFIER_H_
+#define TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_PERMUTATION_PERMUTATION_ARGUMENT_STRINGIFIER_H_
 
 #include <ostream>
 
 #include "tachyon/base/strings/rust_stringifier.h"
-#include "tachyon/zk/plonk/base/column_key_stringifier.h"
+#include "tachyon/zk/plonk/halo2/stringifiers/column_key_stringifier.h"
 #include "tachyon/zk/plonk/permutation/permutation_argument.h"
 
 namespace tachyon::base::internal {
@@ -28,4 +28,4 @@ class RustDebugStringifier<zk::PermutationArgument> {
 
 }  // namespace tachyon::base::internal
 
-#endif  // TACHYON_ZK_PLONK_PERMUTATION_PERMUTATION_ARGUMENT_STRINGIFIER_H_
+#endif  // TACHYON_ZK_PLONK_HALO2_STRINGIFIERS_PERMUTATION_PERMUTATION_ARGUMENT_STRINGIFIER_H_
